rag_app/pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "chromadb>=1.3.2",
    "docling>=2.58.0",
    "fastapi>=0.109.0",
    "flake8>=7.3.0",
    "google-genai>=1.0.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pathlib2>=2.3.7.post1",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.6",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.27.0",
]
